+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -19,6 +19,7 @@ import {perfNow, Job} from '@enact/core/util';
 import {on, off} from '@enact/core/dispatcher';
 import {platform} from '@enact/core/platform';
 import {is} from '@enact/core/keymap';
+import Media from '@enact/ui/Media';
 import Slottable from '@enact/ui/Slottable';
 import Touchable from '@enact/ui/Touchable';
 import Spotlight from '@enact/spotlight';
@@ -59,50 +60,6 @@ const keyMap = {
 	'PAUSE': 19,
 	'REWIND': 412,
 	'FASTFORWARD': 417
-};
-
-// Video ReadyStates
-// - Commented are currently unused.
-//
-// const HAVE_NOTHING      = 0;  // no information whether or not the audio/video is ready
-// const HAVE_METADATA     = 1;  // metadata for the audio/video is ready
-// const HAVE_CURRENT_DATA = 2;  // data for the current playback position is available, but not enough data to play next frame/millisecond
-// const HAVE_FUTURE_DATA  = 3;  // data for the current and at least the next frame is available
-const HAVE_ENOUGH_DATA  = 4;  // enough data available to start playing
-
-
-// Set-up event forwarding map. These are all of the supported media events
-const handledMediaEventsMap = {
-	abort           : 'onAbort',
-	canplay         : 'onCanPlay',
-	canplaythrough  : 'onCanPlayThrough',
-	durationchange  : 'onDurationChange',
-	emptied         : 'onEmptied',
-	encrypted       : 'onEncrypted',
-	ended           : 'onEnded',
-	error           : 'onError',
-	loadeddata      : 'onLoadedData',
-	loadedmetadata  : 'onLoadedMetadata',
-	loadstart       : 'onLoadStart',
-	pause           : 'onPause',
-	play            : 'onPlay',
-	playing         : 'onPlaying',
-	progress        : 'onProgress',
-	ratechange      : 'onRateChange',
-	seeked          : 'onSeeked',
-	seeking         : 'onSeeking',
-	stalled         : 'onStalled',
-	suspend         : 'onSuspend',
-	timeupdate      : 'onTimeUpdate',
-	volumechange    : 'onVolumeChange',
-	waiting         : 'onWaiting'
-};
-
-// List custom events that aren't standard to React. These will be directly added to the video
-// element and props matching their name will be executed as callback functions when the event fires.
-// "umsmediainfo" prop function will execute when the "umsmediainfo" event happens.
-const handledCustomMediaEventsMap = {
-	'umsmediainfo'  : 'onUMSMediaInfo'
 };
 
 // provide forwarding of events on media controls
@@ -592,7 +549,7 @@ const VideoPlayerBase = class extends React.Component {
 
 		/**
 		 * Any children `<source>` tag elements of [VideoPlayer]{@link moonstone/VideoPlayer} will
-		 * be sent directly to the `<video>` element as video sources.
+		 * be sent directly to the `videoComponent` as video sources.
 		 * See: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/source
 		 *
 		 * @type {Node}
@@ -673,7 +630,37 @@ const VideoPlayerBase = class extends React.Component {
 		 * @default 3000
 		 * @public
 		 */
-		tooltipHideDelay: PropTypes.number
+		tooltipHideDelay: PropTypes.number,
+
+		/**
+		 * Video component to use. The default (`'video'`) renders an `HTMLVideoElement`. Custom
+		 * video components must have a similar API structure, exposing the following APIs:
+		 *
+		 * Properties:
+		 * * `currentTime` {Number} - Playback index of the media in seconds
+		 * * `duration` {Number} - Media's entire duration in seconds
+		 * * `error` {Boolean} - `true` if video playback has errored.
+		 * * `loading` {Boolean} - `true` if video playback is loading.
+		 * * `paused` {Boolean} - Playing vs paused state. `true` means the media is paused
+		 * * `playbackRate` {Number} - Current playback rate, as a number
+		 * * `proportionLoaded` {Number} - A value between `0` and `1`
+		 *	representing the proportion of the media that has loaded
+		 * * `proportionPlayed` {Number} - A value between `0` and `1` representing the
+		 *	proportion of the media that has already been shown
+		 *
+		 * Methods:
+		 * * `play()` - play video
+		 * * `pause()` - pause video
+		 * * `load()` - load video
+		 *
+		 * The [`source`]{@link moonstone/VideoPlayer.VideoPlayerBase.source} property is passed to the video
+		 * component as a child node.
+		 *
+		 * @type {Component}
+		 * @default 'video'
+		 * @public
+		 */
+		videoComponent: PropTypes.oneOfType([PropTypes.string, PropTypes.func])
 	}
 
 	static defaultProps = {
@@ -690,7 +677,8 @@ const VideoPlayerBase = class extends React.Component {
 			slowRewind: ['-1/2', '-1']
 		},
 		titleHideDelay: 5000,
-		tooltipHideDelay: 3000
+		tooltipHideDelay: 3000,
+		videoComponent: 'video'
 	}
 
 	constructor (props) {
@@ -698,9 +686,6 @@ const VideoPlayerBase = class extends React.Component {
 
 		// Internal State
 		this.video = null;
-		this.handledMediaForwards = {};
-		this.handledMediaEvents = {};
-		this.handledCustomMediaForwards = {};
 		this.pulsedPlaybackRate = null;
 		this.pulsedPlaybackState = null;
 		this.moreInProgress = false;	// This only has meaning for the time between clicking "more" and the official state is updated. To get "more" state, only look at the state value.
@@ -714,34 +699,17 @@ const VideoPlayerBase = class extends React.Component {
 
 		this.initI18n();
 
-		// Generate event handling forwarders and a smooth block to pass to <Video>
-		for (let key in handledMediaEventsMap) {
-			const eventName = handledMediaEventsMap[key];
-			this.handledMediaForwards[eventName] = forward(eventName);
-			this.handledMediaEvents[eventName] = this.handleEvent;
-		}
-		// Generate event handling forwarders for the custom events too
-		for (let eventName in handledCustomMediaEventsMap) {
-			const propName = handledCustomMediaEventsMap[eventName];
-			const forwardEvent = forward(propName);
-			this.handledCustomMediaForwards[eventName] = ev => forwardEvent(ev, this.props);
-		}
-
 		this.paused = new Pause('VideoPlayer');
 
 		// Re-render-necessary State
 		this.state = {
 			announce: AnnounceState.READY,
-			buffered: 0,
 			currentTime: 0,
 			duration: 0,
 			error: false,
 			loading: false,
-			muted: !!props.muted,
 			paused: props.noAutoPlay,
 			playbackRate: 1,
-			readyState: 0,
-			volume: 1,
 			titleOffsetHeight: 0,
 			bottomOffsetHeight: 0,
 
@@ -768,7 +736,6 @@ const VideoPlayerBase = class extends React.Component {
 		on('keypress', this.activityDetected);
 		on('keydown', this.handleGlobalKeyDown);
 		on('keyup', this.handleKeyUp);
-		this.attachCustomMediaEvents();
 		this.startDelayedFeedbackHide();
 		this.calculateMaxComponentCount();
 	}
@@ -787,7 +754,7 @@ const VideoPlayerBase = class extends React.Component {
 		const {source: nextSource} = nextProps;
 
 		if (!compareSources(source, nextSource)) {
-			this.setState({currentTime: 0, buffered: 0, proportionPlayed: 0, proportionLoaded: 0});
+			this.setState({currentTime: 0, proportionPlayed: 0, proportionLoaded: 0});
 			this.reloadVideo();
 		}
 	}
@@ -805,7 +772,6 @@ const VideoPlayerBase = class extends React.Component {
 			!this.state.mediaSliderVisible && this.state.mediaSliderVisible === nextState.mediaSliderVisible &&
 			this.state.loading === nextState.loading && this.props.loading === nextProps.loading &&
 			(
-				this.state.buffered !== nextState.buffered ||
 				this.state.currentTime !== nextState.currentTime ||
 				this.state.proportionPlayed !== nextState.proportionPlayed ||
 				this.state.sliderTooltipTime !== nextState.sliderTooltipTime
@@ -875,7 +841,6 @@ const VideoPlayerBase = class extends React.Component {
 		off('keypress', this.activityDetected);
 		off('keydown', this.handleGlobalKeyDown);
 		off('keyup', this.handleKeyUp);
-		this.detachCustomMediaEvents();
 		this.stopRewindJob();
 		this.stopAutoCloseTimeout();
 		this.stopDelayedTitleHide();
@@ -928,18 +893,6 @@ const VideoPlayerBase = class extends React.Component {
 			this.locale = locale;
 
 			this.durfmt = new DurationFmt({length: 'medium', style: 'clock', useNative: false});
-		}
-	}
-
-	attachCustomMediaEvents = () => {
-		for (let eventName in this.handledCustomMediaForwards) {
-			on(eventName, this.handledCustomMediaForwards[eventName], this.video);
-		}
-	}
-
-	detachCustomMediaEvents = () => {
-		for (let eventName in this.handledCustomMediaForwards) {
-			off(eventName, this.handledCustomMediaForwards[eventName], this.video);
 		}
 	}
 
@@ -1116,7 +1069,7 @@ const VideoPlayerBase = class extends React.Component {
 			if (this.showMiniFeedback && (!this.state.miniFeedbackVisible || this.state.mediaSliderVisible !== shouldShowSlider)) {
 				this.setState({
 					mediaSliderVisible: shouldShowSlider,
-					miniFeedbackVisible: !(this.state.readyState < HAVE_ENOUGH_DATA || !this.state.duration || this.state.error)
+					miniFeedbackVisible: !(this.state.loading || !this.state.duration || this.state.error)
 				});
 			}
 		}
@@ -1269,21 +1222,17 @@ const VideoPlayerBase = class extends React.Component {
 	updateMainState = () => {
 		const el = this.video;
 		const updatedState = {
-			// Standard video properties
+			// Standard media properties
 			currentTime: el.currentTime,
 			duration: el.duration,
-			buffered: el.buffered,
 			paused: el.playbackRate !== 1 || el.paused,
-			muted: el.muted,
-			volume: el.volume,
 			playbackRate: el.playbackRate,
-			readyState: el.readyState,
 
 			// Non-standard state computed from properties
-			proportionLoaded: el.buffered.length && el.buffered.end(el.buffered.length - 1) / el.duration,
-			proportionPlayed: el.currentTime / el.duration || 0,
-			error: el.networkState === el.NETWORK_NO_SOURCE,
-			loading: el.readyState < el.HAVE_ENOUGH_DATA,
+			proportionLoaded: el.proportionLoaded,
+			proportionPlayed: el.proportionPlayed || 0,
+			error: el.error,
+			loading: el.loading,
 			sliderTooltipTime: this.sliderScrubbing ? (this.sliderKnobProportion * el.duration) : el.currentTime
 		};
 
@@ -1614,14 +1563,17 @@ const VideoPlayerBase = class extends React.Component {
 		this.playbackRate = rate = String(rate);
 		const pbNumber = calcNumberValueOfPlaybackRate(rate);
 
-		// Set native playback rate
-		this.video.playbackRate = pbNumber;
-
 		if (!platform.webos) {
+			// ReactDOM throws error for setting negative value for playbackRate
+			this.video.playbackRate = pbNumber < 0 ? 0 : pbNumber;
+
 			// For supporting cross browser behavior
 			if (pbNumber < 0) {
 				this.beginRewind();
 			}
+		} else {
+			// Set native playback rate
+			this.video.playbackRate = pbNumber;
 		}
 	}
 
@@ -1689,18 +1641,13 @@ const VideoPlayerBase = class extends React.Component {
 
 	handleEvent = (ev) => {
 		this.updateMainState();
+
 		if (ev.type === 'onLoadStart') {
-			this.handleLoadStart(ev);
+			this.handleLoadStart();
 		}
 
 		if (ev.type === 'play') {
 			this.mayRenderBottomControls();
-		}
-		// fetch the forward() we generated earlier, using the event type as a key to find the real event name.
-		const fwd = this.handledMediaForwards[handledMediaEventsMap[ev.type]];
-		if (fwd) {
-			ev = this.addStateToEvent(ev);
-			fwd(ev, this.props);
 		}
 	}
 
@@ -1962,6 +1909,7 @@ const VideoPlayerBase = class extends React.Component {
 			thumbnailComponent,
 			thumbnailSrc,
 			title,
+			videoComponent,
 			...rest} = this.props;
 
 		delete rest.announce;
@@ -1990,11 +1938,6 @@ const VideoPlayerBase = class extends React.Component {
 		delete rest.tooltipHideDelay;
 		delete rest.videoPath;
 
-		// Remove the events we manually added so they aren't added twice or fail.
-		for (let eventName in handledCustomMediaEventsMap) {
-			delete rest[handledCustomMediaEventsMap[eventName]];
-		}
-
 		// Handle some cases when the "more" button is pressed
 		const moreDisabled = !(this.state.more);
 		const controlsAriaProps = this.getControlsAriaProps();
@@ -2010,16 +1953,17 @@ const VideoPlayerBase = class extends React.Component {
 				style={style}
 			>
 				{/* Video Section */}
-				<video
+				<Media
 					{...rest}
 					autoPlay={!noAutoPlay}
 					className={css.video}
+					component={videoComponent}
 					controls={false}
+					onUpdate={this.handleEvent}
 					ref={this.setVideoRef}
-					{...this.handledMediaEvents}
 				>
 					{source}
-				</video>
+				</Media>
 
 				<Overlay
 					bottomControlsVisible={this.state.mediaControlsVisible}